†移動† もともとのテストコードに戻した
