+++ src/test/kotlin/Tester.kt
@@ -1,58 +1,85 @@
 import dev.approvers.jubilant.client.Client
 import dev.approvers.jubilant.client.ClientSettingInfo
-import example.Calculator
-import example.Caller
-import test.EnterEventListener
-import test.Tester
+import dev.approvers.jubilant.commands.CommandResult
+import dev.approvers.jubilant.commands.abc.EventListener
+import dev.approvers.jubilant.commands.abc.PrefixnessCommand
+import dev.approvers.jubilant.commands.annotations.Argument
+import dev.approvers.jubilant.commands.annotations.EventReceiver
+import dev.approvers.jubilant.commands.annotations.SubCommand
+import dev.approvers.jubilant.commands.event.EventType
+import net.dv8tion.jda.api.events.ReadyEvent
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import java.lang.RuntimeException
 
+object Tester : PrefixnessCommand(
+   identify = "test",
+   name = "テスト",
+   description = "さっきのより実用性を高くしました"
+) {
+
+   @SubCommand(identify = "hi", name = "ごあいさつ", description = "ちゃんと挨拶してほら！")
+   fun greeting(args: List<String>, event: MessageReceivedEvent) : CommandResult {
+      event.channel.sendMessage("Hennlo world!").queue()
+      return CommandResult.SUCCESS
+   }
+
+   @SubCommand(identify = "args", name = "引数テスト", description = "引数周り、ちゃんと動いてほしいな")
+   @Argument(count = 3, help = ["引数1", "引数2", "引数3"])
+   fun argument(args: List<String>, event: MessageReceivedEvent) : CommandResult {
+      sendArgsInfo(args, event)
+      return CommandResult.SUCCESS
+   }
+
+   @SubCommand(identify = "args", name = "省略可能な引数がある場合のテスト", description = "引数周り、ちゃんと動いてほしいな")
+   @Argument(count = 3, denyLess = false, help = ["引数1", "引数2", "引数3"])
+   fun lessArgs(args: List<String>, event: MessageReceivedEvent) : CommandResult {
+      sendArgsInfo(args, event)
+      return CommandResult.SUCCESS
+   }
+
+   @SubCommand(identify = "args", name = "いっぱい列挙できる引数が有る場合のテスト", description = "引数周り、ちゃんと動いてほしいな")
+   @Argument(count = 3, denyMore = false, help = ["引数1", "引数2", "引数3"])
+   fun moreArgs(args: List<String>, event: MessageReceivedEvent) : CommandResult {
+      sendArgsInfo(args, event)
+      return CommandResult.SUCCESS
+   }
+
+   private fun sendArgsInfo(args: List<String>, event: MessageReceivedEvent) {
+      event.channel.sendMessage("以下の引数が渡されました:${
+         args.mapIndexed {index: Int, s: String ->
+            "${index + 1} $s"
+         }.joinToString("\n")
+      }")
+   }
+
+}
+
+object EventListener : EventListener() {
+
+   @EventReceiver(EventType.BOT_READY)
+   fun eventReady(event: ReadyEvent) : CommandResult {
+      event.jda.getTextChannelById(695976154779222047)?.sendMessage("Bot ready event dispatched")?.queue()
+      return CommandResult.SUCCESS
+   }
+
+}
 
 fun main() {
 
-    // Botがログを残すチャンネルを取得します。
-    // このプログラムでは環境変数から読み込んでいますが、ハードコーディングでも構いません。
-    val testChannelId = System.getenv("TEST_CHANNEL_ID")?.toLongOrNull()
-        ?: throw RuntimeException("TEST_CHANNEL_ID is not set, or invalid.")
+   val testChannelId = System.getenv("TEST_CHANNEL_ID")?.toLongOrNull()
+      ?: throw RuntimeException("TEST_CHANNEL_ID is not set, or invalid.")
+   val settingInfo = ClientSettingInfo(
+      "Library Test Bot",
+      "[t]:",
+      null,
+      testChannelId,
+      listOf(),
+      true
+   )
 
-    // ClientSettingInfoを作成します。
-    val settingInfo = ClientSettingInfo(
-        // Botの名前。
-        botName = "Jubliant Tester",
+   val client = Client(settingInfo)
+   client.addMessageCommand(Tester)
+   client.addEventListener(EventListener)
+   client.launch()
 
-        // プレフィックス。詳細はREADMEの「コマンドの書式」を確認してください。
-        prefix = "[t]:",
-
-        // トークン。
-        // トークンにnullが渡された場合は、DISCORD_TOKEN環境変数を利用しようとします。
-        // DISCORD_TOKEN環境変数も存在しない場合は、NullPointerExceptionをスローします。
-        token = null,
-
-        // ログ用チャンネルのID。
-        loggingChannelId = testChannelId,
-
-        // Botフィルターから除外するIDのリスト。詳細はREADMEの「動作仕様」を確認してください。
-        botIdsWhiteList = listOf(),
-
-        // ログ用チャンネルのみで動作するか。
-        // falseの場合は、所属するサーバー上の全てのチャンネルで反応します。
-        // trueの場合は、loggingChannelId で指定されたチャンネル上のみで反応します。
-        reactOnlyLoggingChanel = true
-    )
-
-    // 作成したClientSettingInfoからクライアントを生成します。
-    val client = Client(settingInfo)
-
-    // クライアントにコマンドを登録します。
-    client.addMessageCommand(Tester())            // https://github.com/approvers/jubilant/blob/dev/src/test/kotlin/test/Tester.kt
-    client.addMessageCommand(Calculator())        // https://github.com/approvers/jubilant/tree/dev/src/test/kotlin/example/Calculator.kt
-    client.addMessageCommand(Caller())            // https://github.com/approvers/jubilant/tree/dev/src/test/kotlin/example/Caller.kt
-
-    // クライアントにイベントリスナーを登録します。
-    client.addEventListener(EnterEventListener()) // https://github.com/approvers/jubilant/tree/dev/src/test/kotlin/test/Tester.kt
-
-    // クライアントを実行します。
-    // Botの実行が開始されると、別プロセスにフォークされます。
-    client.launch()
-
-    // Botが動作中は、main関数を離脱してもプログラムの実行が停止しません。
 }
